almost done 13/5/24

+++ src/controllers/courseControllers.js
@@ -3,6 +3,8 @@ import { Course } from "../models/courseModel.js";
 import { ErrorHandler } from "../utils/utilityClass.js";
 import cloudinary from "cloudinary";
 import { User } from "../models/userModel.js";
+import { rm } from "fs";
+import { Module } from "../models/moduleModel.js";
 
 export const getAllCourses = async (req, res, next) => {
   try {
@@ -86,17 +88,18 @@ export const getCourseById = async (req, res, next) => {
   }
 };
 
-//<=============================CREATE=====================================>
+//<=============================CREATE=====================================>//
 export const createCourse = async (req, res, next) => {
   try {
     const { subject, Class, description, school } = req.body;
     const { id } = req.query;
     const file = req.file;
-    // console.log(subject, Class, description);
 
-    const fileUri = getDataUri(file);
     const courseExist = await Course.findOne({ subject, class: Class });
     if (courseExist) {
+      rm(file.path, () => {
+        console.log(`${file.originalname} deleted`);
+      });
       return next(
         new ErrorHandler(
           `${courseExist.subject} of class ${courseExist.class} already exist`,
@@ -105,16 +108,12 @@ export const createCourse = async (req, res, next) => {
       );
     }
 
-    const myCloud = await cloudinary.v2.uploader.upload(fileUri.content);
     const course = await Course.create({
       class: Class,
       subject,
       description,
       school,
-      poster: {
-        public_id: myCloud.public_id,
-        url: myCloud.url,
-      },
+      poster: file.path,
       creatorID: id,
     });
 
@@ -131,17 +130,79 @@ export const createCourse = async (req, res, next) => {
 
 export const deleteCourse = async (req, res, next) => {
   try {
-    const id = req.params.id;
-    const course = await Course.findById(id);
-    if (!course) {
-      return next(new ErrorHandler("Course not available", 500));
+    const { cid } = req.params;
+    const { id } = req.query;
+
+    if (!id) {
+      return next(new ErrorHandler("no id", 400));
     }
-    await cloudinary.v2.uploader.destroy(course.poster.public_id);
-    await course.deleteOne();
-    return res.status(200).json({
-      success: true,
-      message: "Course deleted",
-    });
+    const user = await User.findById(id);
+    if (!user) {
+      return next(new ErrorHandler("user not found or invalid id", 404));
+    }
+
+    const course = await Course.findById(cid);
+
+    if (!course) {
+      return next(new ErrorHandler("Course not available", 404));
+    }
+
+    if (user.role === "admin") {
+      rm(course.poster, () => {
+        console.log(`${course.subject}'s poster deleted`);
+      });
+
+      // =====DELETEING ALL MODULES AND THEIR PDFS AND VIDEO LINKS===//
+      for (let i = 0; i < course.modules.length; i++) {
+        let module = await Module.findById(course.modules[i]._id);
+        for (let j = 0; j < module.materials.length; j++) {
+          if (module.materials[j].pname) {
+            rm(module.materials[j].url, () => {
+              console.log(`${module.materials[j].pname}'s poster deleted`);
+            });
+          }
+        }
+        await module.deleteOne();
+      }
+
+      await course.deleteOne();
+      return res.status(200).json({
+        success: true,
+        message: "Course deleted",
+      });
+    } else if (user.role === "teacher") {
+      if (user._id.toString() === course.creatorID.toString()) {
+        rm(course.poster, () => {
+          console.log(`${course.subject}'s poster deleted`);
+        });
+
+        // =====DELETEING ALL MODULES AND THEIR PDFS AND VIDEO LINKS===//
+        for (let i = 0; i < course.modules.length; i++) {
+          let module = await Module.findById(course.modules[i]._id);
+          for (let j = 0; j < module.materials.length; j++) {
+            if (module.materials[j].pname) {
+              rm(module.materials[j].url, () => {
+                console.log(`${module.materials[j].pname}'s poster deleted`);
+              });
+            }
+          }
+          await module.deleteOne();
+        }
+
+        await course.deleteOne();
+        return res.status(200).json({
+          success: true,
+          message: "Course deleted",
+        });
+      }
+      return next(
+        new ErrorHandler("Unauthorized, can't delete other's course", 400)
+      );
+    } else {
+      return next(
+        new ErrorHandler("Unauthorized, can't delete other's course", 400)
+      );
+    }
   } catch (error) {
     console.log(error);
     return next(new ErrorHandler("Error deleting courses ", 500));
